Update Traffic service to BingTraffic. Closes #3417
